Killed dual coordinate systems.
I took out the dual coordinate systems and I made the necessary changes
to methods in the GridMap class. and I simplified the getDryNeighbors()
method in the GridMap class.
I also simplified the getAdjHarePops() and getAdjPumaPops() methods in
the Population class, which both followed a similar structure to the
getDryNeighbors() method.