Changed methods in Population class and TestDriver
I removed some modifications that Sarah put into the Population class.
Instead of having Puma and Hare objects in Population class I want to
have them in the TestDriver class and feed them into the updatePop()
method in the Population class.

I’ve also been working on the TestDriver class and writing it according
to the list of steps that I’ve put forth. I’ve gotten a good ways down
the list so far.